broke out assigned user into separate component
