+++ app/javascript/components/subject/MonitoringActions.js
@@ -1,298 +1,16 @@
 import React from 'react';
 import { PropTypes } from 'prop-types';
-import { Form, Row, Col, Button, Modal } from 'react-bootstrap';
-import _ from 'lodash';
-import axios from 'axios';
-import moment from 'moment';
+import { Form, Row, Col } from 'react-bootstrap';
 
+import AssignedUser from './AssignedUser';
 import CaseStatus from './CaseStatus';
 import Jurisdiction from './Jurisdiction';
 import GenericAction from './GenericAction';
-import DateInput from '../util/DateInput';
 import InfoTooltip from '../util/InfoTooltip';
-import reportError from '../util/ReportError';
 
 class MonitoringActions extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {
-      patient: props.patient,
-      showMonitoringStatusModal: false,
-      showJurisdictionModal: false,
-      showassignedUserModal: false,
-      message: '',
-      reasoning: '',
-      monitoring_status: props.patient.monitoring ? 'Actively Monitoring' : 'Not Monitoring',
-      jurisdiction_path: this.props.jurisdictionPaths[this.props.patient.jurisdiction_id],
-      original_jurisdiction_id: this.props.patient.jurisdiction_id,
-      validJurisdiction: true,
-      assigned_user: props.patient.assigned_user ? props.patient.assigned_user : '',
-      original_assigned_user: props.patient.assigned_user ? props.patient.assigned_user : '',
-      monitoring_reasons: null,
-      monitoring_reason: '',
-      apply_to_group: false,
-      loading: false,
-      apply_to_group_cm_exp_only: false,
-      apply_to_group_cm_exp_only_date: moment(new Date()).format('YYYY-MM-DD'),
-    };
-    this.origState = Object.assign({}, this.state);
-    this.handleChange = this.handleChange.bind(this);
-    this.handleKeyPress = this.handleKeyPress.bind(this);
-    this.submit = this.submit.bind(this);
-    this.toggleMonitoringStatusModal = this.toggleMonitoringStatusModal.bind(this);
-    this.toggleAssignedUserModal = this.toggleAssignedUserModal.bind(this);
-  }
-
-  handleChange(event) {
-    if (event?.target?.name && event.target.name === 'assignedUser') {
-      if (
-        event?.target?.value === '' ||
-        (event?.target?.value && !isNaN(event.target.value) && parseInt(event.target.value) > 0 && parseInt(event.target.value) <= 9999)
-      ) {
-        this.setState({
-          message: `assigned user from "${this.state.original_assigned_user}" to " ${event.target.value}"`,
-          message_warning: '',
-          assigned_user: event?.target?.value ? parseInt(event.target.value) : '',
-          monitoring_reasons: null,
-        });
-      }
-    } else if (event?.target?.id && event.target.id === 'monitoring_status') {
-      this.setState({
-        showMonitoringStatusModal: true,
-        message: `monitoring status to "${event.target.value}"`,
-        message_warning:
-          event.target.value === 'Not Monitoring'
-            ? 'This will move the selected record(s) to the Closed line list and turn Continuous Exposure OFF.'
-            : 'This will move the selected record(s) from the Closed line list to the appropriate Active Monitoring line list',
-        monitoring: event.target.value === 'Actively Monitoring',
-        monitoring_status: event?.target?.value ? event.target.value : '',
-        monitoring_reasons:
-          event.target.value === 'Not Monitoring'
-            ? [
-                'Completed Monitoring',
-                'Meets Case Definition',
-                'Lost to follow-up during monitoring period',
-                'Lost to follow-up (contact never established)',
-                'Transferred to another jurisdiction',
-                'Person Under Investigation (PUI)',
-                'Case confirmed',
-                'Meets criteria to discontinue isolation',
-                'Deceased',
-                'Duplicate',
-                'Other',
-              ]
-            : null,
-      });
-    } else if (event?.target?.name && event.target.name === 'apply_to_group') {
-      let applyToGroup = event.target.id === 'apply_to_group_yes';
-      this.setState({ [event.target.name]: applyToGroup });
-    } else if (event?.target?.name && event.target.name === 'apply_to_group_cm_exp_only') {
-      let applyToGroup = event.target.id === 'apply_to_group_cm_exp_only_yes';
-      this.setState({ [event.target.name]: applyToGroup });
-    } else if (event?.target?.id) {
-      let value = event.target.type === 'checkbox' ? event.target.checked : event.target.value;
-      this.setState({ [event.target.id]: event?.target?.value ? value : '' });
-    }
-  }
-
-  handleKeyPress() {
-    if (event.which === 13) {
-      if (event?.target?.name && event.target.name === 'jurisdictionId') {
-        event.preventDefault();
-        this.toggleJurisdictionModal();
-      } else if (event?.target?.name && event.target.name === 'assignedUser') {
-        event.preventDefault();
-        this.toggleAssignedUserModal();
-      }
-    }
-  }
-
-  toggleMonitoringStatusModal() {
-    let current = this.state.showMonitoringStatusModal;
-    this.setState({
-      showMonitoringStatusModal: !current,
-      monitoring_status: this.props.patient.monitoring ? 'Actively Monitoring' : 'Not Monitoring',
-      apply_to_group: false,
-      reasoning: '',
-    });
-  }
-
-  toggleAssignedUserModal() {
-    let current = this.state.showassignedUserModal;
-    this.setState({
-      message: `assigned user from "${this.state.original_assigned_user}" to "${this.state.assigned_user}"`,
-      showassignedUserModal: !current,
-      assigned_user: current ? this.state.original_assigned_user : this.state.assigned_user,
-      apply_to_group: false,
-      reasoning: '',
-    });
-  }
-
-  submit() {
-    let diffState = Object.keys(this.state).filter(k => _.get(this.state, k) !== _.get(this.origState, k));
-    this.setState({ loading: true }, () => {
-      axios.defaults.headers.common['X-CSRF-Token'] = this.props.authenticity_token;
-      axios
-        .post(window.BASE_PATH + '/patients/' + this.props.patient.id + '/status', {
-          monitoring: this.state.monitoring_status === 'Actively Monitoring',
-          public_health_action: this.state.public_health_action,
-          reasoning:
-            (this.state.showMonitoringStatusModal && this.state.monitoring_status === 'Not Monitoring'
-              ? this.state.monitoring_reason + (this.state.reasoning !== '' ? ', ' : '')
-              : '') + this.state.reasoning,
-          monitoring_reason: this.state.monitoring_status === 'Not Monitoring' ? this.state.monitoring_reason : null,
-          jurisdiction: Object.keys(this.props.jurisdictionPaths).find(id => this.props.jurisdictionPaths[parseInt(id)] === this.state.jurisdiction_path),
-          assigned_user: this.state.assigned_user,
-          apply_to_group: this.state.apply_to_group,
-          isolation: this.state.isolation,
-          pause_notifications: this.state.pause_notifications,
-          diffState: diffState,
-          apply_to_group_cm_exp_only: this.state.apply_to_group_cm_exp_only,
-          apply_to_group_cm_exp_only_date: this.state.apply_to_group_cm_exp_only_date,
-        })
-        .then(() => {
-          if (diffState.includes('jurisdiction_path')) {
-            const currentUserJurisdictionString = this.props.current_user.jurisdiction_path.join(', ');
-            // check if current_user has access to the changed jurisdiction
-            // if so, reload the page, if not, redirect to exposure or isolation dashboard
-            if (!this.state.jurisdiction_path.startsWith(currentUserJurisdictionString)) {
-              const pathEnd = this.state.isolation ? '/isolation' : '';
-              location.assign((window.BASE_PATH ? window.BASE_PATH : '') + '/public_health' + pathEnd);
-            } else {
-              location.reload(true);
-            }
-          } else {
-            location.reload(true);
-          }
-        })
-        .catch(error => {
-          reportError(error);
-        });
-    });
-  }
-
-  createModal(title, toggle, submit) {
-    return (
-      <Modal size="lg" show centered onHide={toggle}>
-        <Modal.Header>
-          <Modal.Title>{title}</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <p>
-            Are you sure you want to change {this.state.message}? {this.state.message_warning && <b>{this.state.message_warning}</b>}
-          </p>
-          {this.props.has_group_members && (
-            <React.Fragment>
-              <p className="mb-2">Please select the records that you would like to apply this change to:</p>
-              <Form.Group className="px-4">
-                <Form.Check
-                  type="radio"
-                  className="mb-1"
-                  name="apply_to_group"
-                  id="apply_to_group_no"
-                  label="This monitoree only"
-                  onChange={this.handleChange}
-                  checked={!this.state.apply_to_group}
-                />
-                <Form.Check
-                  type="radio"
-                  className="mb-3"
-                  name="apply_to_group"
-                  id="apply_to_group_yes"
-                  label="This monitoree and all household members"
-                  onChange={this.handleChange}
-                  checked={this.state.apply_to_group}
-                />
-              </Form.Group>
-              <Form.Group>
-                {(this.state.apply_to_group || this.state.apply_to_group_cm_exp_only) && this.state.household_warning && <i>{this.state.household_warning}</i>}
-              </Form.Group>
-            </React.Fragment>
-          )}
-          {this.state.monitoring_reasons && (
-            <Form.Group>
-              <Form.Label>Please select reason for status change:</Form.Label>
-              <Form.Control as="select" size="lg" className="form-square" id="monitoring_reason" onChange={this.handleChange} defaultValue={-1}>
-                <option value={-1} disabled>
-                  --
-                </option>
-                {this.state.monitoring_reasons.map((option, index) => (
-                  <option key={`option-${index}`} value={option}>
-                    {option}
-                  </option>
-                ))}
-              </Form.Control>
-            </Form.Group>
-          )}
-          <Form.Group>
-            <Form.Label>Please include any additional details:</Form.Label>
-            <Form.Control as="textarea" rows="2" id="reasoning" onChange={this.handleChange} />
-          </Form.Group>
-          {this.props.isolation && this.state.monitoring_reasons && this.props.in_a_group && (
-            <React.Fragment>
-              <hr />
-              <p className="mb-2">
-                Would you like to update the <i>Last Date of Exposure</i> for all household members who have Continuous Exposure turned ON and are being
-                monitored in the Exposure Workflow?
-              </p>
-              <Form.Group className="px-4">
-                <Form.Check
-                  type="radio"
-                  className="mb-2"
-                  name="apply_to_group_cm_exp_only"
-                  id="apply_to_group_cm_exp_only_no"
-                  label="No, household members still have continuous exposure to another case"
-                  onChange={this.handleChange}
-                  checked={!this.state.apply_to_group_cm_exp_only}
-                />
-                <Form.Check>
-                  <Form.Check.Label>
-                    <Form.Check.Input
-                      type="radio"
-                      name="apply_to_group_cm_exp_only"
-                      id="apply_to_group_cm_exp_only_yes"
-                      onChange={this.handleChange}
-                      checked={this.state.apply_to_group_cm_exp_only}
-                    />
-                    <p className="mb-1">Yes, household members are no longer being exposed to a case</p>
-                    {this.state.apply_to_group_cm_exp_only && (
-                      <React.Fragment>
-                        <p className="mb-2">
-                          Update their <b>Last Date of Exposure</b> to:
-                        </p>
-                        <DateInput
-                          id="apply_to_group_cm_exp_only_date"
-                          date={this.state.apply_to_group_cm_exp_only_date}
-                          minDate={'2020-01-01'}
-                          maxDate={moment()
-                            .add(30, 'days')
-                            .format('YYYY-MM-DD')}
-                          onChange={date => this.setState({ apply_to_group_cm_exp_only_date: date })}
-                          placement="bottom"
-                        />
-                      </React.Fragment>
-                    )}
-                  </Form.Check.Label>
-                </Form.Check>
-              </Form.Group>
-            </React.Fragment>
-          )}
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary btn-square" onClick={toggle}>
-            Cancel
-          </Button>
-          <Button variant="primary btn-square" onClick={submit} disabled={this.state.loading}>
-            {this.state.loading && (
-              <React.Fragment>
-                <span className="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>&nbsp;
-              </React.Fragment>
-            )}
-            Submit
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    );
   }
 
   render() {
@@ -307,7 +25,7 @@ class MonitoringActions extends React.Component {
                     MONITORING STATUS
                     <InfoTooltip tooltipTextKey="monitoringStatus" location="right"></InfoTooltip>
                   </Form.Label>
-                  <Form.Control
+                  {/* <Form.Control
                     as="select"
                     className="form-control-lg"
                     id="monitoring_status"
@@ -315,7 +33,7 @@ class MonitoringActions extends React.Component {
                     value={this.state.monitoring_status}>
                     <option>Actively Monitoring</option>
                     <option>Not Monitoring</option>
-                  </Form.Control>
+                  </Form.Control> */}
                 </Form.Group>
                 <Form.Group as={Col} md="12" lg="8" className="pt-2">
                   <GenericAction
@@ -364,40 +82,12 @@ class MonitoringActions extends React.Component {
                   />
                 </Form.Group>
                 <Form.Group as={Col} md="12" lg="8" className="pt-2">
-                  <Form.Label className="nav-input-label">
-                    ASSIGNED USER
-                    <InfoTooltip tooltipTextKey="assignedUser" location="right"></InfoTooltip>
-                  </Form.Label>
-                  <Form.Group className="d-flex mb-0">
-                    <Form.Control
-                      as="input"
-                      name="assignedUser"
-                      list="assignedUsers"
-                      autoComplete="off"
-                      className="form-control-lg"
-                      onChange={this.handleChange}
-                      onKeyPress={this.handleKeyPress}
-                      value={this.state.assigned_user}
-                    />
-                    <datalist id="assignedUsers">
-                      {this.props.assignedUsers.map(num => {
-                        return (
-                          <option value={num} key={num}>
-                            {num}
-                          </option>
-                        );
-                      })}
-                    </datalist>
-                    {this.state.assigned_user === this.state.original_assigned_user ? (
-                      <Button className="btn-lg btn-square text-nowrap ml-2" disabled>
-                        <i className="fas fa-users"></i> Change User
-                      </Button>
-                    ) : (
-                      <Button className="btn-lg btn-square text-nowrap ml-2" onClick={this.toggleAssignedUserModal}>
-                        <i className="fas fa-users"></i> Change User
-                      </Button>
-                    )}
-                  </Form.Group>
+                  <AssignedUser
+                    patient={this.props.patient}
+                    authenticity_token={this.props.authenticity_token}
+                    has_group_members={this.props.has_group_members}
+                    assignedUsers={this.props.assignedUsers}
+                  />
                 </Form.Group>
                 <Form.Group as={Col} lg="24" className="pt-2">
                   <Jurisdiction
@@ -412,8 +102,6 @@ class MonitoringActions extends React.Component {
             </Col>
           </Row>
         </Form>
-        {this.state.showMonitoringStatusModal && this.createModal('Monitoring Status', this.toggleMonitoringStatusModal, this.submit)}
-        {this.state.showassignedUserModal && this.createModal('Assigned User', this.toggleAssignedUserModal, this.submit)}
       </React.Fragment>
     );
   }
